Refactors email sender configuration
Replaces the use of `DefaultSenderEmail` and `DefaultSenderName` with a list of `EmailSender` objects in the email provider configuration.

This change allows users to define multiple senders per integration, instead of just one default sender. It also fixes a bug where an invalid UUID sender ID in email provider settings would cause an error. Now, the invalid UUID will be automatically replaced.

The template validation is updated to check for the presence of `sender_id` instead of `from_address`.